+++ src/main/java/library/service/UserService.java
@@ -2,11 +2,15 @@ package library.service;
 
 import library.common.UserRole;
 import library.controller.DTO.UserDTO.GetUserFullDTO;
+import library.exception.PendingReturn;
 import library.exception.UserNotFoundForId;
 import library.exception.UserNotFoundForUsername;
 import library.infrastructure.entity.AuthEntity;
+import library.infrastructure.entity.RentalEntity;
 import library.infrastructure.entity.UserEntity;
 import library.infrastructure.repository.AuthRepository;
+import library.infrastructure.repository.RentalRepository;
+import library.infrastructure.repository.ReviewRepository;
 import library.infrastructure.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -19,11 +23,16 @@ import java.util.List;
 public class UserService {
     private final UserRepository userRepository;
     private final AuthRepository authRepository;
+    private final RentalRepository rentalRepository;
+    private final ReviewRepository reviewRepository;
 
     @Autowired
-    public UserService(UserRepository userRepository, AuthRepository authRepository) {
+    public UserService(UserRepository userRepository, AuthRepository authRepository, RentalRepository rentalRepository,
+                       ReviewRepository reviewRepository) {
         this.userRepository = userRepository;
         this.authRepository = authRepository;
+        this.rentalRepository = rentalRepository;
+        this.reviewRepository = reviewRepository;
     }
 
     public List<GetUserFullDTO> getAll() {
@@ -53,8 +62,26 @@ public class UserService {
         if (!userRepository.existsById(id)) {
             throw UserNotFoundForId.create(id);
         }
-        authRepository.deleteByUserId(id);
-        userRepository.deleteById(id);
+        boolean noPendingReturns = false;
+        if (!rentalRepository.findByUserId(id).isEmpty()) {
+            List<RentalEntity> rentals = rentalRepository.findByUserId(id);
+            for (RentalEntity rental : rentals) {
+                if (rental.getReturnDate() != null) {
+                    noPendingReturns = true;
+                } else {
+                    throw PendingReturn.create(rental.getBook().getTitle());
+                }
+            }
+        } else {
+            authRepository.deleteByUserId(id);
+            userRepository.deleteById(id);
+        }
+        if (noPendingReturns) {
+            rentalRepository.deleteByUserId(id);
+            reviewRepository.deleteByUserId(id);
+            authRepository.deleteByUserId(id);
+            userRepository.deleteById(id);
+        }
     }
 
     private GetUserFullDTO mapUser(UserEntity user, AuthEntity auth) {
